requirements.txt
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
tzdata==2025.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2